ESP32 platform
This is enough for all the test app features to work on ESP32 without
SSL.

+++ lib/handshake.c
@@ -245,7 +245,7 @@ postbody_completion:
 
 read_ok:
 	/* Nothing more to do for now */
-	lwsl_info("%s: read_ok, used %ld\n", __func__, buf - oldbuf);
+	lwsl_info("%s: read_ok, used %ld\n", __func__, (long)(buf - oldbuf));
 
 	return buf - oldbuf;
